Matrix Profile transformer added
